Moves setMembersOnly method to ChatRoom.
